Rename prismic.js
